fix all warnings

+++ plasm2_emu/cpu/ops/cops_devices.c
@@ -10,7 +10,7 @@ plasm2_emu
 
 void DSQ(void) {
 	byte Register = mmu_read1(i->ip++) & 0xF;
-	u64 Status = devicesi_statusquery(i->rs_gprs[Register] & 0xFFFFFFFF);
+	u64 Status = devicesi_statusquery((u32)i->rs_gprs[Register] & 0xFFFFFFFF);
 	mmu_push(Status);
 	return;
 }
@@ -24,7 +24,7 @@ void DSC(void) {
 		};
 	}Inputs;
 	Inputs.Input = mmu_read1(i->ip++);
-	devicesi_sendcommand(i->rs_gprs[Inputs.Device], i->rs_gprs[Inputs.Command]);
+	devicesi_sendcommand((u32)i->rs_gprs[Inputs.Device], i->rs_gprs[Inputs.Command]);
 	return;
 }
 
@@ -37,7 +37,7 @@ void DSD(void) {
 		};
 	}Inputs;
 	Inputs.Input = mmu_read1(i->ip++);
-	devicesi_senddata(i->rs_gprs[Inputs.Device], i->rs_gprs[Inputs.Data]);
+	devicesi_senddata((u32)i->rs_gprs[Inputs.Device], i->rs_gprs[Inputs.Data]);
 	return;
 }
 
@@ -50,25 +50,25 @@ void DGD(void) {
 		};
 	}Inputs;
 	Inputs.Input = mmu_read1(i->ip++);
-	i->rs_gprs[Inputs.Destination] = devicesi_getdata(i->rs_gprs[Inputs.Device]);
+	i->rs_gprs[Inputs.Destination] = devicesi_getdata((u32)i->rs_gprs[Inputs.Device]);
 	return;
 }
 
 void DRS(void) {
 	u64 Device = i->rs_gprs[mmu_read1(i->ip++) & 0xF];
-	devicesi_reset(Device);
+	devicesi_reset((u32)Device);
 	return;
 }
 
 void DPE(void) {
 	u64 Device = i->rs_gprs[mmu_read1(i->ip++) & 0xF];
-	devicesi_on(Device);
+	devicesi_on((u32)Device);
 	return;
 }
 
 void DPD(void) {
 	u64 Device = i->rs_gprs[mmu_read1(i->ip++) & 0xF];
-	devicesi_off(Device);
+	devicesi_off((u32)Device);
 	return;
 }
